+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+* Fix minification issue from previous release ([#648](https://github.com/evanw/esbuild/issues/648))
+
+    The minification optimization to omit certain `continue` and `return` statements when it's implied by control flow in version 0.8.29 caused a regression when the branch condition uses a hoisted function:
+
+    ```js
+    if (fn()) return;
+    ...
+    function fn() {}
+    ```
+
+    In that case, transforming the code by inverting the condition and moving the following statements inside the branch is not valid because the function is no longer hoisted to above the branch condition. This release fixes the regression by avoiding this optimization in cases like this.
+
 * Tree-shake unused code with `--format=iife` ([#639](https://github.com/evanw/esbuild/issues/639))
 
     When the output format is IIFE (which wraps the code in an immediately-invoked function expression), esbuild now assumes that it's safe to remove unused code. This is an assumption that esbuild always makes when bundling but that esbuild previously didn't make when not bundling. Now esbuild will remove code even when not bundling as long as the output format is IIFE.